style: move tests into colocation
